moved the cell based interpolant to seperate class
